Backend/requirement.txt
tensorflow
numpy
opencv-python
flask
pillow
matplotlib
